requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
pydantic>=2.0.0
